safe area added to appbar , padding update
